+++ R/model.R
@@ -1,6 +1,6 @@
 #' @include build.R
 #' @include run.R
-#' @include load.R
+#' @include loadOutput.R
 NULL
 
 #' Model
@@ -119,7 +119,7 @@ model <- function(
   m1 = 0.035,
   CO2 = 360,
   # species data
-  species = read.table(getOption("RconTroll.path"), header=TRUE, dec=".", sep=""),
+  species = read.table(getOption("RconTroll.species"), header=TRUE, dec=".", sep=""),
   # climate data
   Tyear = c(24.64982014,	24.60624211,	24.49933474,	24.96279395,	24.88365139,
             24.87594184,	24.98313802,	25.68910135,	26.58033805,	26.98619405,
@@ -214,7 +214,7 @@ model <- function(
       overwrite = overwrite,
       verbose = verbose)
   path <- file.path(path, name)
-  model <- load(name = name, path = path)
+  model <- loadOutput(name = name, path = path)
 
   return(model)
 }