+++ R/print.TROLLsimstack.R
@@ -3,7 +3,7 @@ NULL
 
 #' Function to print TROLL outputs.
 #'
-#' @param x TROLLsim
+#' @param x TROLLsimstack
 #'
 #' @return Print in console
 #'
@@ -11,34 +11,31 @@ NULL
 #'
 #' @examples
 #'
-#' @name print.TROLLsim
+#' @name print.TROLLsimstack
 #' 
-setMethod('print', 'TROLLsim', function(x, ...) {
+setMethod('print', 'TROLLsimstack', function(x, ...) {
 
   cat('Object of class :', class(x)[1],'\n\n')
-
-  cat('Name :', x@name, '\n\n')
-
-  cat('2D discrete network: horizontal step = ', x@info$step, 'm, one tree per 1 m^2 \n')
-  cat('Number of sites      : ', x@info$SitesNb[1], 'x', x@info$SitesNb[2], '\n')
-  cat('Number of iterations : ', x@info$IterationsNb, '\n')
-  cat('Duration of timestep : ', x@info$timestep, 'years \n')
-  cat('Number of Species    : ', x@info$SpeciesNb, '\n\n')
-
-  cat('Average computation time : ', x@info$ComputationTime, 'seconds \n\n')
+  
+  cat('Structured :', x@structured, '\n')
+  cat('Compressed', x@compressed, '\n\n')
+  
+  if(isTRUE(x@structured)){
+  cat('Number of sites      : ', x@nbcols, 'x', x@nbrows, '\n')
+  cat('Number of iterations : ', x@nbiter, '\n')
+  cat('Duration of timestep : ', 1/x@iter, 'years \n')}
+  
 })
 
-setMethod('show', 'TROLLsim', function(object) {
+setMethod('show', 'TROLLsimstack', function(object) {
 
   cat('Object of class :', class(object)[1],'\n\n')
-
-  cat('Name :', object@name, '\n\n')
-
-  cat('2D discrete network: horizontal step = ', object@info$step, 'm, one tree per 1 m^2 \n')
-  cat('Number of sites      : ', object@info$SitesNb[1], 'object', object@info$SitesNb[2], '\n')
-  cat('Number of iterations : ', object@info$IterationsNb, '\n')
-  cat('Duration of timestep : ', object@info$timestep, 'years \n')
-  cat('Number of Species    : ', object@info$SpeciesNb, '\n\n')
-
-  cat('Average computation time : ', object@info$ComputationTime, 'seconds \n\n')
+  
+  cat('Structured :', object@structured, '\n')
+  cat('Compressed', object@compressed, '\n\n')
+  
+  if(isTRUE(object@structured)){
+    cat('Number of sites      : ', object@nbcols, 'x', object@nbrows, '\n')
+    cat('Number of iterations : ', object@nbiter, '\n')
+    cat('Duration of timestep : ', 1/object@iter, 'years \n')}
 })