+++ NEWS.md
@@ -1,9 +1,17 @@
 # TO DO
-
+* modification of input file
+* merging of info and par file in output, potentially using info-file for module activation
+* finish `summary` for TROLLsimstack
+* create subclasses for TROLLsimstack
+* define `plot` for TROLLsimstack
+* Correct `plot.output` method with use of internal functions (see disturbance example)
+* Integrate virtual forest in modelling process (activate _from _FromData in `build` and integrate forest.txt in `run`)
 
 # RconTroll 0.1.9003
 
-* renaming objects and methods (e.g. `TROLLoutput` to `TROLLsim`, `virtualizeFromData` to `inventoryFromData`)
+* renaming objects and methods (e.g. `TROLLoutput` to `TROLLsim`, `virtualizeFromData` to `inventoryFromData`, etc.)
+* definition of `summary` for class TROLLsim
+* new class TROLLsimstack (with methods `print` and `summary`, the latter not yet fully implemented)
 
 # TO DO